fix(build): resolve all MSVC warnings across src/, examples/, tests/
- C4244/C4267 (narrowing): static_cast in LQA, FS1052 ARQ, examples, tests
- C4996 (deprecated str*): #pragma warning(disable:4996) in files with
  safe strncpy/strcpy patterns (fixed-size ALE address fields)
- C4100 (unreferenced param): comment out unused param names
- C4189 (unused local var): remove or suppress unused variables
- C4804 (bool in >=): fix count>=2 → words.size()>=2 in test_address_encoder
- C4127 (constant condition): suppress in test_fsk_core (intentional runtime checks)

Clean build: 0 warnings. All 31 tests pass.

+++ examples/example_ale_decoder.cpp
@@ -65,8 +65,8 @@ void example_individual_call() {
     
     FFTDemodulator demodulator;
     
-    auto to_detected = demodulator.process_audio(to_audio.data(), to_audio.size());
-    auto from_detected = demodulator.process_audio(from_audio.data(), from_audio.size());
+    auto to_detected = demodulator.process_audio(to_audio.data(), static_cast<uint32_t>(to_audio.size()));
+    auto from_detected = demodulator.process_audio(from_audio.data(), static_cast<uint32_t>(from_audio.size()));
     
     std::cout << "Demodulated: " << to_detected.size() << " + " 
               << from_detected.size() << " symbols\n";
